web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.120.3"
uvicorn = "^0.38.0"
python-multipart = "^0.0.20"
openai = "^2.6.1"
aiofiles = "^25.1.0"
jinja2 = "^3.1.6"
python-dotenv = "^1.2.1"
python-docx = "^1.2.0"
pdfplumber = "^0.11.7"
python-pptx = "^1.0.2"
redis = "^7.0.1"
rq = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
